store: Implement a new MatrixStore.
The old store used the user ID as the primary key for the accounts table
but used the device ID as the foreign key for the other tables (Olm
sessions, inbound group sessions).

While it is very unlikely this could end up leaking sessions between
differing user ids provided that they get the same device ID.

The new store uses a normal auto-incrementing integer value as the
primary key and automatically updates from the old schema to the new
one.