events: Remove the RoomEncryptedEvent.
The RoomEncryptedEvent class was added to represent m.room.encrypted
events.

Olm events as well as megolm events use that event type and the
algorithm needs to be checked to differentiate them.

The parsing path that is set up this way allows both m.room.encrypted
event types in the to-device sync part as well as in the room timelines.

Since Olm events aren't supported in the room timeliens and megolm
events in to-device messages this leads to alowing unexpeced events in
differing parts of a sync response.

Separate those two out so they can only appear in their repsective
places.