crypto: Remove the key query TODO for unknown devices when resharing keys.
If an unknown device that doesn't belong to our own user requests a key
such it is guaranteed that the key wasn't shared with the device in the
first place.

Since we only share keys that were already shared with third party
devices it's safe to ignore such a request without further action.