Correctly handle server errors for uploads
If a server error occurs (timeout or 429), we can only discard the data
uploaded until now and start over.

upload() now takes a function that returns data instead of data
directly, it will be called initially and on retries to get new data
and start over. This allows the client user to do any cleanup they need,
seek(0, 0) a file to return it again, etc.