Reduce default AsyncClient request timeout
Adds an option for AsyncClientConfig to specify a timeout for aiohttp
ClientSession requests, defaulting to 30 seconds instead of 5 minutes.

`sync()` will override this with `timeout` parameter + 5 seconds (to
give the server a chance to return normally).
This notably fixes an issue where `sync_forever()` seems to hang
eternally after a machine wakes up from hibernation or
network interface changes even though a `timeout` argument value
is passed.

`download()`, `thumbnail()` and `upload()` will ignore this and pass a
timeout of 0 (no timeout), since we don't want to kill user
download/uploads.