tests: Fix the sync forever test.
This test used to fail because the sync response would end up reporting
a bigger key uploaded key count than the keys upload response.

This would end up in a race where the sync claims that no keys need to be
uploaded but the upload would claim that more keys need to be uplaoded.