Refactor _build_path to accept path as List[str]
Previously, accepting `path` as a "str" caused a bug
where valid matrix names (like "@a-z0-9._=-/:example.com")
were breaking as the "/" was interpreted as a path separation.

The benefit of moving to a List[str] is that each substring of the path
can be separetly percent-encoded and then combined into one long
path string. The above username becomes
"%40a-z0-9._%3D-%2F%3Aexample.com" as there is no confusion
whether that is a name or a path.