fix(sync): pass the default request timeout to the API
When calling either `sync` or `sync_forever` methods without passing the
`timeout` parameter, the value used for requests timeout is supposed to
be read from the client config.

While this value is correctly passed to the `ClientSession`, it is not
included as a parameter in the URL of the `/sync` API call at all,
making call return immediately, often with an empty response.

This change makes the API call use the `config.request_timeout` value
instead if the sync timeout was explicitely not set. The value is cast
as an integer and converted to milliseconds to match what's expected
by the API.