fix: Element exports keys without required fields
When trying to import files from an Element backup, some of the exported keys are missing the "sender_claimed_keys" dictionary key. These entries in the encrypted file have the following shape:

```
{
  'sender_key': '... ~256 bytes ...',
  'room_id': '!room:homeserver',
  'session_id': '... ~256 bytes ...',
  'session_key': '... long value ...',
  'forwarding_curve25519_key_chain': [],
  'org.matrix.msc3061.shared_history': False,
  'algorithm': 'm.megolm.v1.aes-sha2'
}
```

As a result, importing the whole file fails because of the schema validation for `Schemas.megolm_key_import`. This patch silently continues, but warns the user that some of the keys in the file weren't decrypted. I'm not familiar enough with the details of megolm; but this patch correctly works around this issue, instead of hard failing. I hope it is useful for someone!