Enable chunked downloads
Before this change, whenever the user would download an attachment, the
client (Pantalaimon) would store all the data in memory while processing
it. This causes significant memory usage spikes when downloading
multiple large files at once. The high memory usage is a major problem
when running Pantalaimon in a memory-constrained setting.

This change adds a function for lazily decrypting files (an analogue to
the one for lazily encrypting files, which already exists), and a new
client function for using the chunked data.