Fix type of call event version (#521)
Currently, the schema of the call events specifies the field "version"
as integer. The matrix specification [1] lists the type of the field as
both string and integer (old spec version). Also, receiving and handling
call events on a server like matrix.org does not work and results in the
following log messages:

WARNI Error validating event: '1' is not of type 'integer'

Failed validating 'type' in
schema['properties']['content']['properties']['version']:
    {'type': 'integer'}

On instance['content']['version']:
    '1'

To fix this issue, change the type from integer to integer and string
and update the tests accordingly.

[1]: https://spec.matrix.org/v1.11/client-server-api/#common-fields

Signed-off-by: hwipl <33433250+hwipl@users.noreply.github.com>